+++ upload/admin/model/extension/payment/g2apay.php
@@ -1,6 +1,5 @@
 <?php
 class ModelExtensionPaymentG2apay extends Model {
-
 	public function install() {
 		$this->db->query("
 			CREATE TABLE `" . DB_PREFIX . "g2apay_order` (
@@ -41,6 +40,7 @@ class ModelExtensionPaymentG2apay extends Model {
 		if ($qry->num_rows) {
 			$order = $qry->row;
 			$order['transactions'] = $this->getTransactions($order['g2apay_order_id'], $qry->row['currency_code']);
+
 			return $order;
 		} else {
 			return false;
@@ -50,7 +50,7 @@ class ModelExtensionPaymentG2apay extends Model {
 	public function getTotalReleased($g2apay_order_id) {
 		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "g2apay_order_transaction` WHERE `g2apay_order_id` = '" . (int)$g2apay_order_id . "' AND (`type` = 'payment' OR `type` = 'refund')");
 
-		return (double)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 
 	public function refund($g2apay_order, $amount) {
@@ -69,7 +69,7 @@ class ModelExtensionPaymentG2apay extends Model {
 			$fields = array(
 				'action' => 'refund',
 				'amount' => $refunded_amount,
-				'hash' => $hash,
+				'hash'   => $hash,
 			);
 
 			return $this->sendCurl($url, $fields);
@@ -91,6 +91,7 @@ class ModelExtensionPaymentG2apay extends Model {
 				$row['amount'] = $this->currency->format($row['amount'], $currency_code, true, true);
 				$transactions[] = $row;
 			}
+
 			return $transactions;
 		} else {
 			return false;
@@ -98,13 +99,13 @@ class ModelExtensionPaymentG2apay extends Model {
 	}
 
 	public function addTransaction($g2apay_order_id, $type, $total) {
-		$this->db->query("INSERT INTO `" . DB_PREFIX . "g2apay_order_transaction` SET `g2apay_order_id` = '" . (int)$g2apay_order_id . "',`date_added` = now(), `type` = '" . $this->db->escape($type) . "', `amount` = '" . (double)$total . "'");
+		$this->db->query("INSERT INTO `" . DB_PREFIX . "g2apay_order_transaction` SET `g2apay_order_id` = '" . (int)$g2apay_order_id . "',`date_added` = now(), `type` = '" . $this->db->escape($type) . "', `amount` = '" . (float)$total . "'");
 	}
 
 	public function getTotalRefunded($g2apay_order_id) {
 		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "g2apay_order_transaction` WHERE `g2apay_order_id` = '" . (int)$g2apay_order_id . "' AND 'refund'");
 
-		return (double)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 
 	public function sendCurl($url, $fields) {
@@ -117,9 +118,11 @@ class ModelExtensionPaymentG2apay extends Model {
 		$auth_hash = hash('sha256', $this->config->get('payment_g2apay_api_hash') . $this->config->get('payment_g2apay_username') . html_entity_decode($this->config->get('payment_g2apay_secret')));
 		$authorization = $this->config->get('payment_g2apay_api_hash') . ";" . $auth_hash;
 		curl_setopt(
-				$curl, CURLOPT_HTTPHEADER, array(
-			"Authorization: " . $authorization
-				)
+			$curl,
+			CURLOPT_HTTPHEADER,
+			array(
+				"Authorization: " . $authorization
+			)
 		);
 
 		$response = json_decode(curl_exec($curl));
