+++ upload/admin/model/extension/payment/globalpay_remote.php
@@ -7,7 +7,7 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 			return "https://api.realexpayments.com/epage-remote.cgi";
 		}
 	}
-	
+
 	public function install() {
 		$this->db->query("
 			CREATE TABLE IF NOT EXISTS `" . DB_PREFIX . "globalpay_remote_order` (
@@ -75,8 +75,8 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 			curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
 			curl_setopt($ch, CURLOPT_POSTFIELDS, $xml);
 			curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
-			$response = curl_exec ($ch);
-			curl_close ($ch);
+			$response = curl_exec($ch);
+			curl_close($ch);
 
 			return simplexml_load_string($response);
 		} else {
@@ -97,26 +97,26 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 			$secret = $this->config->get('payment_globalpay_remote_secret');
 
 			if ($globalpay_order['settle_type'] == 2) {
-				$this->logger('Capture hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount*100) . '.' . (string)$globalpay_order['currency_code'] . '.');
+				$this->logger('Capture hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount * 100) . '.' . (string)$globalpay_order['currency_code'] . '.');
 
-				$tmp = $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount*100) . '.' . (string)$globalpay_order['currency_code'] . '.';
+				$tmp = $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount * 100) . '.' . (string)$globalpay_order['currency_code'] . '.';
 				$hash = sha1($tmp);
 				$tmp = $hash . '.' . $secret;
 				$hash = sha1($tmp);
 
 				$settle_type = 'multisettle';
-				$xml_amount = '<amount currency="' . (string)$globalpay_order['currency_code'] . '">' . (int)round($amount*100) . '</amount>';
+				$xml_amount = '<amount currency="' . (string)$globalpay_order['currency_code'] . '">' . (int)round($amount * 100) . '</amount>';
 			} else {
 				//$this->logger('Capture hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '...');
-				$this->logger('Capture hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount*100) . '.' . (string)$globalpay_order['currency_code'] . '.');
+				$this->logger('Capture hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount * 100) . '.' . (string)$globalpay_order['currency_code'] . '.');
 
-				$tmp = $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount*100) . '.' . (string)$globalpay_order['currency_code'] . '.';
+				$tmp = $timestamp . '.' . $merchant_id . '.' . $globalpay_order['order_ref'] . '.' . (int)round($amount * 100) . '.' . (string)$globalpay_order['currency_code'] . '.';
 				$hash = sha1($tmp);
 				$tmp = $hash . '.' . $secret;
 				$hash = sha1($tmp);
 
 				$settle_type = 'settle';
-				$xml_amount = '<amount currency="' . (string)$globalpay_order['currency_code'] . '">' . (int)round($amount*100) . '</amount>';
+				$xml_amount = '<amount currency="' . (string)$globalpay_order['currency_code'] . '">' . (int)round($amount * 100) . '</amount>';
 			}
 
 			$xml = '';
@@ -139,8 +139,8 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 			curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
 			curl_setopt($ch, CURLOPT_POSTFIELDS, $xml);
 			curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
-			$response = curl_exec ($ch);
-			curl_close ($ch);
+			$response = curl_exec($ch);
+			curl_close($ch);
 
 			return simplexml_load_string($response);
 		} else {
@@ -177,9 +177,9 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 				$pas_ref = $globalpay_order['pasref'];
 			}
 
-			$this->logger('Rebate hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $order_ref . '.' . (int)round($amount*100) . '.' . $globalpay_order['currency_code'] . '.');
+			$this->logger('Rebate hash construct: ' . $timestamp . '.' . $merchant_id . '.' . $order_ref . '.' . (int)round($amount * 100) . '.' . $globalpay_order['currency_code'] . '.');
 
-			$tmp = $timestamp . '.' . $merchant_id . '.' . $order_ref . '.' . (int)round($amount*100) . '.' . $globalpay_order['currency_code'] . '.';
+			$tmp = $timestamp . '.' . $merchant_id . '.' . $order_ref . '.' . (int)round($amount * 100) . '.' . $globalpay_order['currency_code'] . '.';
 			$hash = sha1($tmp);
 			$tmp = $hash . '.' . $secret;
 			$hash = sha1($tmp);
@@ -193,7 +193,7 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 			$xml .= '<orderid>' . $order_ref . '</orderid>';
 			$xml .= '<pasref>' . $pas_ref . '</pasref>';
 			$xml .= '<authcode>' . $globalpay_order['authcode'] . '</authcode>';
-			$xml .= '<amount currency="' . (string)$globalpay_order['currency_code'] . '">' . (int)round($amount*100) . '</amount>';
+			$xml .= '<amount currency="' . (string)$globalpay_order['currency_code'] . '">' . (int)round($amount * 100) . '</amount>';
 			$xml .= '<refundhash>' . $rebatehash . '</refundhash>';
 			$xml .= '<sha1hash>' . $hash . '</sha1hash>';
 			$xml .= '</request>';
@@ -207,8 +207,8 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 			curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
 			curl_setopt($ch, CURLOPT_POSTFIELDS, $xml);
 			curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
-			$response = curl_exec ($ch);
-			curl_close ($ch);
+			$response = curl_exec($ch);
+			curl_close($ch);
 
 			return simplexml_load_string($response);
 		} else {
@@ -263,6 +263,6 @@ class ModelExtensionPaymentGlobalpayRemote extends Model {
 	public function getTotalRebated($globalpay_order_id) {
 		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "globalpay_remote_order_transaction` WHERE `globalpay_remote_order_id` = '" . (int)$globalpay_order_id . "' AND 'rebate'");
 
-		return (double)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 }
